chore: rename l2tol2msg package to interop
